backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
supabase==2.5.1
transformers==4.41.2
torch==2.3.1
feedparser==6.0.11
beautifulsoup4==4.12.3
pytest==8.2.2
python-dotenv==1.0.1
httpx==0.27.0
